Change Description :
HotfixUtil.java: added a -1 to the split method so that it doesn't
				ignore the empty splits.
Index.html : Made multiple changes on index.html to handle the UI 			
			 exceptions and placed the divs accordingly. Completely formatted the
			 code for beautification purposes.
Readme.txt : Added few changes to Readme.txt.
ResultGenerator : Identified a bug, where there is a multiple entry for
the same record, added the break statements accordingly.
Once all the records are added, the while condition evaluates to true
again.. breaking that.